requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
reportlab==4.2.5
